Modificata initfs_open Aggiornata la open (ora assegna al campo fs_spec_id di fd_list il valore ritornato dalla initfs_open Aggiunto Makefile.am per dare una compilazione piu pulita. Eliminati alcuni warning.

+++ shell/commands.c
@@ -422,7 +422,7 @@ void tester(){
 		return;
 	} else { 
 		for ( i = 0 ; i <= MAX_TEST ; i++) {
-			if ( (strcmp(argv[1], testing[i].cmd_testname) ) == NULL ) {
+			if ( (strcmp(argv[1], testing[i].cmd_testname) ) == (int) NULL ) {
 				 (testing[i].func)();
 				break;
 			}
